[ECP-8992] Create abstract config route patcher
