+++ Setup/Patch/Abstract/AbstractConfigurationPathPatcher.php
@@ -1,26 +1,55 @@
 <?php
+/**
+ *
+ * Adyen ExpressCheckout Module
+ *
+ * Copyright (c) 2024 Adyen N.V.
+ * This file is open source and available under the MIT license.
+ * See the LICENSE file for more info.
+ *
+ * Author: Adyen <magento@adyen.com>
+ */
 declare(strict_types=1);
 
-namespace Adyen\ExpressCheckout\Helper;
+namespace Adyen\ExpressCheckout\Setup\Patch\Abstract;
 
 use Magento\Framework\App\Config\ReinitableConfigInterface;
 use Magento\Framework\App\Config\Storage\WriterInterface;
 use Magento\Framework\Setup\ModuleDataSetupInterface;
+use Magento\Framework\Setup\Patch\DataPatchInterface;
 
-class DataPatch
+abstract class AbstractConfigurationPathPatcher implements DataPatchInterface
 {
+    private ModuleDataSetupInterface $moduleDataSetup;
     private WriterInterface $configWriter;
     private ReinitableConfigInterface $reinitableConfig;
 
     public function __construct(
+        ModuleDataSetupInterface $moduleDataSetup,
         WriterInterface $configWriter,
         ReinitableConfigInterface $reinitableConfig
     ) {
+        $this->moduleDataSetup = $moduleDataSetup;
         $this->configWriter = $configWriter;
         $this->reinitableConfig = $reinitableConfig;
     }
 
-    public function updateConfigValue(
+    public function apply(): void
+    {
+        $this->moduleDataSetup->getConnection()->startSetup();
+
+        foreach (static::REPLACE_CONFIG_PATHS as $oldConfigPath => $newConfigPath) {
+            $this->updateConfigValue(
+                $this->moduleDataSetup,
+                $oldConfigPath,
+                $newConfigPath
+            );
+        }
+
+        $this->moduleDataSetup->getConnection()->endSetup();
+    }
+
+    private function updateConfigValue(
         ModuleDataSetupInterface $setup,
         string $oldPath,
         string $newPath
@@ -53,5 +82,21 @@ class DataPatch
 
         $matchingConfigs = $connection->fetchAll($select);
         return reset($matchingConfigs);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getAliases(): array
+    {
+        return [];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public static function getDependencies(): array
+    {
+        return [];
     }
 }
