+++ Setup/Patch/Data/UpdateExpressDBPaths.php
@@ -13,64 +13,14 @@ declare(strict_types=1);
 
 namespace Adyen\ExpressCheckout\Setup\Patch\Data;
 
-use Adyen\ExpressCheckout\Helper\DataPatch;
-use Magento\Framework\Setup\ModuleDataSetupInterface;
+use Adyen\ExpressCheckout\Setup\Patch\Abstract\AbstractConfigurationPathPatcher;
 use Magento\Framework\Setup\Patch\DataPatchInterface;
 
-class UpdateExpressDBPaths implements DataPatchInterface
+class UpdateExpressDBPaths extends AbstractConfigurationPathPatcher implements DataPatchInterface
 {
-    private ModuleDataSetupInterface $moduleDataSetup;
-    private DataPatch $dataPatchHelper;
-
-    public function __construct(
-        ModuleDataSetupInterface $moduleDataSetup,
-        DataPatch $dataPatchHelper
-    ) {
-        $this->moduleDataSetup = $moduleDataSetup;
-        $this->dataPatchHelper = $dataPatchHelper;
-    }
-
-    public function apply(): void
-    {
-        $this->moduleDataSetup->getConnection()->startSetup();
-
-        // Update Apple Pay config path
-        $this->dataPatchHelper->updateConfigValue(
-            $this->moduleDataSetup,
-            'payment/adyen_hpp/show_apple_pay_on',
-            'payment/adyen_express/show_apple_pay_on'
-        );
-
-        // Update Google Pay config path
-        $this->dataPatchHelper->updateConfigValue(
-            $this->moduleDataSetup,
-            'payment/adyen_hpp/show_google_pay_on',
-            'payment/adyen_express/show_google_pay_on'
-        );
-
-        // Update Google Pay config path
-        $this->dataPatchHelper->updateConfigValue(
-            $this->moduleDataSetup,
-            'payment/adyen_hpp/apple_pay_button_color',
-            'payment/adyen_express/apple_pay_button_color'
-        );
-
-        $this->moduleDataSetup->getConnection()->endSetup();
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function getAliases(): array
-    {
-        return [];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public static function getDependencies(): array
-    {
-        return [];
-    }
+    const REPLACE_CONFIG_PATHS = [
+        'payment/adyen_hpp/show_apple_pay_on' => 'payment/adyen_express/show_apple_pay_on',
+        'payment/adyen_hpp/show_google_pay_on' => 'payment/adyen_express/show_google_pay_on',
+        'payment/adyen_hpp/apple_pay_button_color' => 'payment/adyen_express/apple_pay_button_color'
+    ];
 }
